test(cluster): split distributor support coverage
